requirements-dev.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
basemap=1.0.8.dev0=np110py27_1
basemap-data-hires=1.0.8.dev0=0
blas=1.1=openblas
curl=7.45.0=0
cycler=0.10.0=py27_0
execnet=1.2.0=py27_0
fontconfig=2.11.1=2
freetype=2.6.3=0
functools32=3.2.3.2=py27_1
geopy=1.10.0=py27_0
geos=3.4.2=1
hdf5=1.8.15.1=2
icu=56.1=2
jpeg=9b=0
libgcc=5.2.0=0
libgfortran=3.0=0
libiconv=1.14=1
libnetcdf=4.3.3.1=3
libpng=1.6.17=0
libtiff=4.0.6=1
libxml2=2.9.2=0
libxslt=1.1.28=0
lxml=3.6.0=py27_0
matplotlib=1.5.1=np110py27_1
mkl=11.3.1=0
netcdf4=1.2.4=np110py27_1
numpy=1.10.4=py27_blas_openblas_201
openblas=0.2.18=1
openssl=1.0.2g=0
pep8=1.7.0=py27_0
pillow=3.2.0=py27_2
pip=8.1.1=py27_1
py=1.4.31=py27_0
pyparsing=2.0.3=py27_0
pyproj=1.9.5.1=py27_0
pyqt=4.11.4=py27_1
pyshp=1.2.3=py27_0
pytest=2.9.1=py27_0
pytest-cache=1.0=py27_0
pytest-pep8=1.0.6=py27_0
python=2.7.11=0
python-dateutil=2.5.2=py27_0
pytz=2016.3=py27_0
qt=4.8.7=1
readline=6.2=2
requests=2.9.1=py27_0
scipy=0.17.0=np110py27_2
setuptools=21.2.1=py27_0
simplejson=3.8.2=py27_0
sip=4.16.9=py27_0
six=1.10.0=py27_0
sqlite=3.13.0=0
tk=8.5.19=0
wheel=0.29.0=py27_0
xz=5.2.2=0
zlib=1.2.8=0
